srotasHealth/clinical-ai/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
google-genai
pypdf
python-multipart
faiss-cpu
sentence-transformers
pydantic